requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pillow==10.0.0
streamlit==1.28.0
locust==2.17.0
fastapi==0.110.0
uvicorn==0.29.0
pydantic
pandas
scikit-learn
python-multipart
